Refrigerator_Temparature_Control_&_Alerting_System/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
python-dotenv==1.1.1